requirements.txt
Django==5.0.6
djangorestframework==3.15.2
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.135.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytz==2024.1
Pillow==10.4.0
